ai_code/mcq/requirement.txt
streamlit
langchain_core
langchain_groq
langchain_community
Wikipedia-API
duckduckgo-search
youtube-transcript-api
chromadb
reportlab 
markdown 
beautifulsoup4
langchain
pypdf
python-dotenv
ddgs
fpdf
openai
langchain-openai
langchain-google-genai